crypto: change type uint32_t to size_t in aes_cypher() arguments.
VELAPLATFO-1905

+++ arch/arm/src/lpc43xx/lpc43_aes.c
@@ -161,8 +161,8 @@ static int aes_update(FAR const void *out,
  * Public Functions
  ****************************************************************************/
 
-int aes_cypher(void *out, const void *in, uint32_t size, const void *iv,
-               const void *key, uint32_t keysize, int mode, int encrypt)
+int aes_cypher(void *out, const void *in, size_t size, const void *iv,
+               const void *key, size_t keysize, int mode, int encrypt)
 {
   unsigned int ret = 0;
   uint32_t outl = size;
